plc v2.0.0.2

+++ _third_part/algorithm/audio/process/plc/musicplc/music_plc_algo_api_c.cpp
@@ -7,10 +7,10 @@ void plc_register_music_plc_int16() {
 	plc_register_c::Register<music_plc_algo_api_c<int16_t>>("mplc_i16");
 }
 void plc_register_music_plc_int32() {
-	plc_register_c::Register<music_plc_algo_api_c<int16_t>>("mplc_i32");
+	plc_register_c::Register<music_plc_algo_api_c<int32_t>>("mplc_i32");
 }
 void plc_register_music_plc_intf32() {
-	plc_register_c::Register<music_plc_algo_api_c<int16_t>>("mplc_if32");
+	plc_register_c::Register<music_plc_algo_api_c<f32>>("mplc_if32");
 }
 
 }
